Split discovery checks in 2 tests & create PR with new API changes  - One is bound to PR/merges to ensure the discovery file is valid  - Other is used daily to provide a summary of the API validations errors
