+++ .github/workflows/sync-discovery.yml
@@ -1,0 +1,33 @@
+# This workflow will review the current status of the APIs. It will verify if the API is:
+#  - Reachable, i.e. we can download it.
+#  - Following OpenapiV3+
+
+name: Sync discovery
+
+on:
+  schedule:
+    - cron: "0 0 * * *"
+
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [18.x]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Use Node.js ${{ matrix.node-version }}
+      uses: actions/setup-node@v3
+      with:
+        node-version: ${{ matrix.node-version }}
+        cache: 'npm'
+    - run: npm ci
+    - run: npm run discovery
+    - name: Create Pull Request
+      uses: peter-evans/create-pull-request@v4
+      with:
+        title: Syncing API
+        commit-message: Syncing API
+        delete-branch: true
