+++ .github/workflows/validate-discovery-file.yml
@@ -1,0 +1,29 @@
+# This workflow will review the current status of the APIs. It will verify if the API is:
+#  - Reachable, i.e. we can download it.
+#  - Following OpenapiV3+
+
+name: Validate discovery file
+
+on:
+  pull_request:
+    branches: [ "main" ]
+    paths:
+      - 'discovery/**'
+
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [18.x]
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Use Node.js ${{ matrix.node-version }}
+      uses: actions/setup-node@v3
+      with:
+        node-version: ${{ matrix.node-version }}
+        cache: 'npm'
+    - run: npm ci
+    - run: npm exec jest --src discovery/Discovery.test.ts
