+++ discovery/Discovery.test.ts
@@ -1,6 +1,5 @@
 import {App, Discovery, getPath, Tag} from "./Discovery";
-import SwaggerParser from '@apidevtools/swagger-parser';
-import {readFileSync} from 'fs';
+import {readFileSync, existsSync} from 'fs';
 import path from 'path';
 import {parse} from 'yaml';
 import Ajv from 'ajv';
@@ -29,16 +28,6 @@ const publicApiFilter = (app: App): boolean => !skipFilter(app) && !app.useLocal
 const privateApiFilter = (app: App): boolean => !skipFilter(app) && !!app.useLocalFile;
 const idMapper = (withId: WithId): string => withId.id;
 
-const validApiTest = async (_ignore: string, app: App, groupId: string) => {
-    const api = await SwaggerParser.validate(getPath(app, groupId));
-    expect(api).toBeTruthy();
-
-    expect(api).toHaveProperty('openapi');
-
-    const version = (api as any).openapi as string;
-    expect(version).toMatch(/^3(.\d(.\d)?)?/);
-}
-
 const expectNonRepeatedIds = (arrayOfItemsWithId: Array<WithId>) => {
     const idSet = new Set(arrayOfItemsWithId.map(idMapper));
     const repeatedIds = [];
@@ -53,7 +42,7 @@ const expectNonRepeatedIds = (arrayOfItemsWithId: Array<WithId>) => {
     expect(repeatedIds).toHaveLength(0);
 }
 
-describe('Discovered OpenAPI are valid OpenAPI3.0+', () => {
+describe('Discovery', () => {
 
     test('Discovery file is valid', () => {
         const validate = new Ajv().compile(DiscoverySchema);
@@ -70,10 +59,6 @@ describe('Discovered OpenAPI are valid OpenAPI3.0+', () => {
 
     describe.each(Object.values(discovery.apis).map(g => [`${g.name}(${g.id})`, g.apps, g.id]))('%s', (_ignore, apps, group) => {
         const prepareApp = (app: App) => [`${app.name}(${app.id})`, app, group] as const;
-
-        const skipped = apps.filter(skipFilter);
-
-        const appsWithPublicUrls = apps.filter(publicApiFilter);
         const appsWithNonPublicUrls = apps.filter(privateApiFilter);
 
         test('No repeated app.id within groups', () => {
@@ -90,34 +75,12 @@ describe('Discovered OpenAPI are valid OpenAPI3.0+', () => {
             }
         })
 
-        test('Testing all APIs one way or another', () => {
-            const allPieces = [
-                ...skipped.map(idMapper),
-                ...appsWithPublicUrls.map(idMapper),
-                ...appsWithNonPublicUrls.map(idMapper)
-            ];
-
-            expect(allPieces).toHaveLength(apps.length);
-            expect(allPieces).toEqual(expect.arrayContaining(apps.map(idMapper)));
-        });
-
-        if (appsWithPublicUrls.length > 0) {
-            describe('public url', () => {
-                test.each(appsWithPublicUrls.map(prepareApp))('%s', validApiTest);
-            });
-        }
-
         if (appsWithNonPublicUrls.length > 0) {
             describe('local file', () => {
-                test.each(appsWithNonPublicUrls.map(prepareApp))('%s', validApiTest);
-            });
-        }
-        if (skipped.length > 0) {
-            describe('Skipped', () => {
-                test.skip.each(skipped.map(a => `${a.name}(${a.id}) - ${skipReason(a)}`))('%s', () => {
-                    throw new Error('Should skip this test');
+                test.each(appsWithNonPublicUrls.map(prepareApp))('%s', (_ignore: string, app: App, groupId: string) => {
+                    expect(existsSync(getPath(app, groupId))).toBeTruthy();
                 });
-            })
+            });
         }
     })
 });
